parserutils: extract shared PopulateErrorDetails into new package
This allows us to break the dependency from `jsonpath/parser` on
`sql/parser`.

Additionally, it fixes the bug where we used the hardcoded (inherited
from `sql/parser`) code for ERROR token in other parsers - the method
now explicitly takes that code as an argument.

Release note: None
