+++ pkg/BUILD.bazel
@@ -771,6 +771,7 @@ ALL_TESTS = [
     "//pkg/util/json/tokenizer:tokenizer_test",
     "//pkg/util/json:json_disallowed_imports_test",
     "//pkg/util/json:json_test",
+    "//pkg/util/jsonpath/parser:parser_disallowed_imports_test",
     "//pkg/util/jsonpath/parser:parser_test",
     "//pkg/util/limit:limit_test",
     "//pkg/util/log/eventlog:eventlog_test",
@@ -2164,6 +2165,7 @@ GO_TARGETS = [
     "//pkg/sql/parser/statements:statements",
     "//pkg/sql/parser:parser",
     "//pkg/sql/parser:parser_test",
+    "//pkg/sql/parserutils:parserutils",
     "//pkg/sql/pgrepl/lsn:lsn",
     "//pkg/sql/pgrepl/lsnutil:lsnutil",
     "//pkg/sql/pgrepl/pgreplparser:pgreplparser",
