+++ pkg/sql/parser/BUILD.bazel
@@ -26,6 +26,7 @@ go_library(
         "//pkg/security/username",  # keep
         "//pkg/sql/lexbase",
         "//pkg/sql/parser/statements",
+        "//pkg/sql/parserutils",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/privilege",  # keep
