+++ pkg/sql/pgrepl/pgreplparser/BUILD.bazel
@@ -37,8 +37,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/sql/lexbase",
-        "//pkg/sql/parser",
         "//pkg/sql/parser/statements",
+        "//pkg/sql/parserutils",
         "//pkg/sql/pgrepl/lsn",
         "//pkg/sql/pgrepl/pgrepltree",
         "//pkg/sql/pgwire/pgcode",
