+++ pkg/sql/pgrepl/pgreplparser/testdata/parser/start_replication.ddt
@@ -30,7 +30,7 @@ expected a positive integer for timeline
 parse error
 START_REPLICATION SLOT slot_1 PHYSICAL A0010/1011 TIMELINE 3354731204578932752358923785325325
 ----
-at or near "3354731204578932752358923785325325": syntax error
+lexical error: 3354731204578932752358923785325325
 DETAIL: source SQL:
 START_REPLICATION SLOT slot_1 PHYSICAL A0010/1011 TIMELINE 3354731204578932752358923785325325
                                                            ^
