+++ pkg/sql/pgrepl/pgreplparser/testdata/parser/timeline_history.ddt
@@ -26,7 +26,7 @@ expected a positive integer for timeline
 parse error
 TIMELINE_HISTORY 133032475823895789234759378957238957328943289745381974981238947
 ----
-at or near "133032475823895789234759378957238957328943289745381974981238947": syntax error
+lexical error: 133032475823895789234759378957238957328943289745381974981238947
 DETAIL: source SQL:
 TIMELINE_HISTORY 133032475823895789234759378957238957328943289745381974981238947
                  ^
