+++ pkg/sql/plpgsql/parser/BUILD.bazel
@@ -37,6 +37,7 @@ go_library(
         "//pkg/build",  # keep
         "//pkg/sql/parser",
         "//pkg/sql/parser/statements",
+        "//pkg/sql/parserutils",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/plpgsql/parser/lexbase",  # keep
